src/requirements.txt
matplotlib==3.7.2
pandas==2.0.3
plotly==5.9.0
seaborn==0.13.0
streamlit==1.34.0
google-generativeai
langchain-google-genai
python-dotenv
pydantic-settings==2.2.1